updating repo with changes we made in class today. changes are primarily focused on routing with handlebars in viewRoutes
